[FIX] error codes and implements test for check correctness

+++ src/main/java/com/zzpj/backend/api/v1/UserController.java
@@ -4,6 +4,7 @@ import com.zzpj.backend.entities.User;
 import com.zzpj.backend.services.interfaceses.UserServiceLocal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,43 +22,46 @@ public class UserController {
     }
 
     @GetMapping
-    public List<User> getAll(){
+    public List<User> getAll() {
         return userService.getAllUsers();
     }
 
     @GetMapping("{id}")
-    public User get(@PathVariable Long id){
+    public User get(@PathVariable Long id) {
         Optional<User> user = userService.getUser(id);
         return user.orElseGet(User::new);
     }
 
     @PostMapping
-    public HttpStatus add(@RequestBody User user){
+    public ResponseEntity add(@RequestBody User user) {
         try {
-        userService.addUser(user);
-    } catch (Exception e) {
-        return HttpStatus.INTERNAL_SERVER_ERROR;
-    }
-        return HttpStatus.CREATED;
+            if(user.getId() != null && user.getId() < 0) throw new Exception();
+            userService.addUser(user);
+        } catch (Exception e) {
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity(HttpStatus.CREATED);
     }
 
     @PutMapping
-    public HttpStatus edit(@RequestBody User user){
+    public ResponseEntity edit(@RequestBody User user) {
         try {
+            if(user.getId() != null && user.getId() < 0) throw new Exception();
             userService.editUser(user);
-    } catch (Exception e) {
-        return HttpStatus.NOT_MODIFIED;
-    }
-        return HttpStatus.OK;
+        } catch (Exception e) {
+            return new ResponseEntity(HttpStatus.NOT_MODIFIED);
+        }
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     @DeleteMapping("{id}")
-    public HttpStatus delete(@PathVariable Long id){
-       try {
-           userService.deleteUser(id);
-    } catch (Exception e) {
-        return HttpStatus.INTERNAL_SERVER_ERROR;
-    }
-        return HttpStatus.OK;
+    public ResponseEntity delete(@PathVariable Long id) {
+        try {
+            if(id < 0) throw new Exception();
+            userService.deleteUser(id);
+        } catch (Exception e) {
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity(HttpStatus.OK);
     }
 }
